debian: add directories "available" in triggers
If one of these files change, an nginx reload will be required since
it might be symlinked into the corresponding "enabled" directory.

Also, this is the most simple solutions to fix the fact the trigger is
not "fired" when only symbolic link are created in the "enabled"
directory. We still the triggers on the "enabled" directories in the
case a package ships a file directly into it (altough it's not
recommended to do so).