debian: don't fail on nginx reload failure
This is a bit of a poor man solution to the problem of upgrades in XiVO,
where quite a few package dependencies are not explicit and the upgrade
is thus fragile.

In this case, what was happening during a 15.08 to 16.13 upgrade is:

1. xivo-config was upgraded and its postinst removed the certificates in
   /etc/nginx/ssl
2. xivo-confd in version 15.04 was still installed and the nginx
   configuration it shipped had a dependency on the presence of the
   certificates in /etc/nginx/ssl (thus, there should have been a "Breaks"
   dependency in xivo-config for xivo-confd)
3. xivo-nginx was installed and did a "service nginx reload", but the
   xivo-confd nginx configuration was invalid (referencing inexisting
   certificates) and thus the reload failed, and thus the xivo-nginx
   package installation failed

That said, not failing in the postinst when nginx reload fails seems the
right thing to do since xivo-nginx doesn't control the whole nginx
configuration and the package installation should not fail even when the
nginx configuration is broken.