throttling: add a noauth zone
The noauth zone should be used on any resource that is not authenticated. The
limit is by IP and a maximum of 10Mb is used for the state of this zone.

The rate limit is 15 request/seconds. Futher requests will receive a 429
response.

The "wazo-no-auth-shared.conf" rules can be included in any location using the
include directive.

include /etc/nginx/wazo-no-auth-shared.conf;