Move cors file handling to a file that needs to be included
I could not get the version in the server block to work. One of the reason is
that add add_header directives disappear once we do a add_header in the
location.

I also introduced a directory named all-locations.d that is used to include some
logic that should replicated to all locations.

All service that want to benefit from this change will need to

include /etc/nginx/locations/https-available/all-locations.d/*;